backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-openai==0.3.0
langchain-pinecone==0.2.0
pinecone-client==5.0.1
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
xmltodict==0.14.1
python-multipart==0.0.20
slowapi==0.1.9
ragas==0.2.14
datasets==3.2.0
numpy
rank-bm25==0.2.2
flashrank==0.2.7
